fix(workflows): repair the merge gate the v2 migration broke
Repo shape: split WorkflowBuilderSurface into the surface and its inspector
pane (selection resolution moves to lib/domain, where components/** must put
non-UI logic), and split the two builder suites that crossed the 600-line cap
along their own seams.

Release gate: the shipped starter template this ladder renamed to
"Bug investigation" left T3-WF-1's structural copy pinning the old slug, so
its drift guard failed. The copy now names the shipped template.

Tier-2 gate: the Blank workflow item grew a "⌘N" badge, which both changed
the item's accessible name out from under the definition-lifecycle spec and
named a keystroke bound app-wide to workspace.new-default (open new chat).
Dropped the badge and pinned the exact accessible name in the unit suite.
