+++ apps/packages/product-client/src/components/workflows/builder-v2/WorkflowBuilderSurface.test.tsx
@@ -381,278 +381,6 @@ describe("WorkflowBuilderSurface", () => {
     expect(onSaved).toHaveBeenCalledWith("wf-new");
   });
 
-  it("shows an unresolved reference on the step that spells it and blocks Save", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    fireEvent.change(screen.getByLabelText("Prompt"), {
-      target: { value: "Write @doc:findings." },
-    });
-
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-    const alerts = screen.getAllByRole("alert").map((node) => node.textContent ?? "");
-    expect(alerts.some((text) => text.includes("@doc:findings"))).toBe(true);
-  });
-
-  it("offers the catalog's models once a harness is picked", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    expect(screen.getByLabelText("Model")).toHaveProperty("disabled", true);
-    fireEvent.change(screen.getByLabelText("Harness"), { target: { value: "claude" } });
-
-    const model = screen.getByLabelText("Model");
-    expect(model).toHaveProperty("disabled", false);
-    expect([...(model as HTMLSelectElement).options].map((option) => option.value))
-      .toEqual(["", "sonnet", "opus"]);
-  });
-
-  it("keeps the model pick when a step is toggled to require approval", async () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    completeSelectedStep();
-    fireEvent.change(screen.getByLabelText("Harness"), { target: { value: "claude" } });
-    fireEvent.change(screen.getByLabelText("Model"), { target: { value: "opus" } });
-
-    fireEvent.click(screen.getByLabelText("Requires human approval"));
-
-    // A gated step still runs an agent session, so the model configuration
-    // survives the toggle: the fields stay rendered, editable, and populated.
-    expect(screen.queryByText("The run pauses here until someone approves the step.")).toBeNull();
-    expect(screen.getByLabelText("Harness")).toHaveProperty("disabled", false);
-    expect(screen.getByLabelText("Model")).toHaveProperty("value", "opus");
-
-    fireEvent.click(screen.getByRole("button", { name: "Save Workflow" }));
-    await waitFor(() => expect(mocks.create).toHaveBeenCalledTimes(1));
-    // The wire shape is unchanged by design: the saved node carries both the
-    // gate and the model.
-    expect(mocks.create.mock.calls[0][0].definition.nodes[0]).toMatchObject({
-      type: "human_in_loop",
-      model: { agentKind: "claude", modelId: "opus" },
-    });
-
-    // Toggling back off keeps it too — the toggle only moves `type`.
-    fireEvent.click(screen.getByLabelText("Requires human approval"));
-    expect(screen.getByLabelText("Model")).toHaveProperty("value", "opus");
-  });
-
-  it("chips a malformed prompt reference and blocks Save", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    fireEvent.change(screen.getByLabelText("Title"), { target: { value: "Diagnose" } });
-    // `@doc:plan.md` is well-formed prose and a malformed reference: declaring
-    // a document cannot rescue it, so it must block the save on its own.
-    fireEvent.change(screen.getByLabelText("Prompt"), {
-      target: { value: "Read @doc:plan.md first." },
-    });
-
-    expect(screen.getByText("@doc:plan.md").getAttribute("data-malformed")).toBe("true");
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-    const alerts = screen.getAllByRole("alert").map((node) => node.textContent ?? "");
-    expect(alerts.some((text) => text.includes("malformed reference “@doc:plan.md”"))).toBe(true);
-
-    // Negative control: the same prompt with a well-formed slug saves once the
-    // document is declared.
-    fireEvent.change(screen.getByLabelText("Prompt"), {
-      target: { value: "Read @doc:plan first." },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "Add document" }));
-    fireEvent.change(screen.getByLabelText("Slug"), { target: { value: "plan" } });
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", false);
-  });
-
-  it("marks a node id the grammar refuses on the id itself", () => {
-    // Ids are minted valid, so this can only arrive from a definition authored
-    // elsewhere — the case the badge has to be honest about.
-    mocks.detailQuery.data = {
-      ...createdRecord(),
-      definition: {
-        schemaVersion: 2,
-        nodes: [{ id: "2-step", type: "agent", title: "Diagnose", prompt: "" }],
-        edges: [],
-        inputs: [],
-        docTemplates: [],
-      },
-    };
-    render(
-      <WorkflowBuilderSurface
-        definitionId="wf-new"
-        authCacheScope="user-1"
-      />,
-    );
-
-    expect(screen.getByText("2-step").getAttribute("data-invalid")).toBe("true");
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Renamed" } });
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-  });
-
-  it("reports a declared input name against its own row and blocks Save", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    completeSelectedStep();
-    fireEvent.click(screen.getByRole("button", { name: /Trigger payload/ }));
-    fireEvent.click(screen.getByRole("button", { name: "Add input" }));
-    fireEvent.change(screen.getByLabelText("Name"), { target: { value: "my input" } });
-
-    // The value the author typed is kept verbatim — a silent rename would
-    // change what every prompt has to spell — and the field carries the error.
-    expect(screen.getByLabelText("Name")).toHaveProperty("value", "my input");
-    expect(screen.getByLabelText("Name").getAttribute("aria-invalid")).toBe("true");
-    const alerts = screen.getAllByRole("alert").map((node) => node.textContent ?? "");
-    expect(alerts.some((text) => text.includes("Input name “my input” must start with a letter")))
-      .toBe(true);
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-
-    // Negative control: one grammatical name away from savable.
-    fireEvent.change(screen.getByLabelText("Name"), { target: { value: "my_input" } });
-    expect(screen.getByLabelText("Name").getAttribute("aria-invalid")).toBeNull();
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", false);
-  });
-
-  it("lowercases a doc slug as it is typed and reports what it cannot rescue", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    completeSelectedStep();
-    fireEvent.click(screen.getByRole("button", { name: "Add document" }));
-    fireEvent.change(screen.getByLabelText("Slug"), { target: { value: " Research-Findings " } });
-
-    expect(screen.getByLabelText("Slug")).toHaveProperty("value", "research-findings");
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", false);
-
-    fireEvent.change(screen.getByLabelText("Slug"), { target: { value: "My_Doc" } });
-
-    expect(screen.getByLabelText("Slug")).toHaveProperty("value", "my_doc");
-    expect(screen.getByLabelText("Slug").getAttribute("aria-invalid")).toBe("true");
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-  });
-
-  it("offers the runtime's repo roots as the default repository", () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.click(screen.getByRole("button", { name: /Trigger payload/ }));
-    const picker = screen.getByLabelText("Default repository");
-    expect([...(picker as HTMLSelectElement).options].map((option) => [
-      option.value,
-      option.textContent,
-    ])).toEqual([
-      ["", "Ask at launch"],
-      ["root-1", "proliferate"],
-      ["root-2", "sidecar"],
-    ]);
-  });
-
-  it("saves the picked repo root as the definition's default", async () => {
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    completeSelectedStep();
-    fireEvent.click(screen.getByRole("button", { name: /Trigger payload/ }));
-    fireEvent.change(screen.getByLabelText("Default repository"), {
-      target: { value: "root-2" },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "Save Workflow" }));
-
-    await waitFor(() => expect(mocks.create).toHaveBeenCalledTimes(1));
-    // The runtime repo-root id travels verbatim: the trigger dialog seeds
-    // `placement.repoConfigId` from this field.
-    expect(mocks.create.mock.calls[0][0].defaultRepoConfigId).toBe("root-2");
-  });
-
-  it("blocks Save while the saved default is not a listed repo root", () => {
-    mocks.detailQuery.data = { ...createdRecord(), defaultRepoConfigId: "repo-config-9" };
-    render(
-      <WorkflowBuilderSurface
-        definitionId="wf-new"
-        authCacheScope="user-1"
-      />,
-    );
-
-    // Edit something first: an untouched record has nothing to save either way.
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Renamed" } });
-
-    fireEvent.click(screen.getByRole("button", { name: /Trigger payload/ }));
-    expect(screen.getByText("Saved repository unavailable (repo-config-9)")).toBeTruthy();
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", true);
-
-    // Negative control: the same edit saves once a listed root is picked.
-    fireEvent.change(screen.getByLabelText("Default repository"), {
-      target: { value: "root-1" },
-    });
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", false);
-  });
-
-  it("says so when the runtime's repositories cannot be loaded", () => {
-    mocks.repoRootsQuery.data = undefined;
-    mocks.repoRootsQuery.isError = true;
-    render(
-      <WorkflowBuilderSurface
-        definitionId={null}
-        template={null}
-        authCacheScope="user-1"
-      />,
-    );
-
-    expect(screen.getByText(
-      "Repositories could not be loaded from the runtime. Reconnect to change this workflow's "
-      + "default repository.",
-    )).toBeTruthy();
-    // A workflow that names no repository is still savable: the run picks one.
-    fireEvent.change(screen.getByLabelText("Workflow title"), { target: { value: "Issue triage" } });
-    completeSelectedStep();
-    expect(screen.getByRole("button", { name: "Save Workflow" })).toHaveProperty("disabled", false);
-  });
-
   it("refuses to open a definition that is not schema version 2", () => {
     mocks.detailQuery.data = {
       ...createdRecord(),
